+++ ChangeLog.md
@@ -1,6 +1,10 @@
+### 0.12.0 / 2019-11-23
+* Raise ParseError if parse+convert convenience method (to_pitch, to_meter, etc.) fails
+* Simplify Key class
+* Add parsing support for key signatures
+
 ### 0.11.2 / 2016-03-29
 * Fix a bug in SuperCollider performance code
-* Raise ParseError if parse+convert convenience method (to_pitch, to_meter, etc.) fails
 
 ### 0.11.1 / 2016-02-18
 * Fix gem description string in .gemspec